refactor: インターフェースのエクスポートを削除し、型定義を簡素化
