+++ packages/ui/src/components/shogi-match.tsx
@@ -6,7 +6,6 @@ import {
     createDefaultNnueSelection,
     createEmptyHands,
     DEFAULT_PRESET_KEY,
-    deriveLastMove,
     type GameResult,
     getAllSquares,
     getPositionService,
@@ -44,6 +43,7 @@ import { useBatchAnalysis } from "./shogi-match/hooks/useBatchAnalysis";
 import { type ClockSettings, useClockManager } from "./shogi-match/hooks/useClockManager";
 import { useEngineManager } from "./shogi-match/hooks/useEngineManager";
 import { useEnginePool } from "./shogi-match/hooks/useEnginePool";
+import { useKifuImportExport } from "./shogi-match/hooks/useKifuImportExport";
 import { useKifuKeyboardNavigation } from "./shogi-match/hooks/useKifuKeyboardNavigation";
 import { useKifuNavigation } from "./shogi-match/hooks/useKifuNavigation";
 import { useLocalStorage } from "./shogi-match/hooks/useLocalStorage";
@@ -72,8 +72,7 @@ import {
     countPieces,
 } from "./shogi-match/utils/boardUtils";
 import { isPromotable, PIECE_CAP, PIECE_LABELS } from "./shogi-match/utils/constants";
-import { exportToKifString, type KifMove } from "./shogi-match/utils/kifFormat";
-import { type KifMoveData, parseSfen } from "./shogi-match/utils/kifParser";
+import type { KifMove } from "./shogi-match/utils/kifFormat";
 import { LegalMoveCache } from "./shogi-match/utils/legalMoveCache";
 import {
     buildPassRightsOptionForLegalMoves,
@@ -1742,82 +1741,27 @@ export function ShogiMatch({
         moveProcessingRef,
     });
 
-    const loadMoves = useCallback(
-        async (
-            list: string[],
-            moveData: KifMoveData[] | undefined,
-            startPosition: PositionState,
-            startSfenToLoad: string,
-        ) => {
-            const filtered = list.filter(Boolean);
-            const service = getPositionService();
-            // パス入り棋譜の場合はpassRightsを渡す
-            const passRightsOption = buildPassRightsOptionForLegalMoves(
-                passRightsSettings,
-                filtered,
-            );
-            const result = await service.replayMovesStrict(
-                startSfenToLoad,
-                filtered,
-                passRightsOption,
-            );
-
-            // 棋譜ナビゲーションをリセット
-            navigation.reset(startPosition, startSfenToLoad);
-            setLastAddedBranchInfo(null); // 分岐状態をクリア
-
-            // 各手を順番に追加
-            let currentPos = startPosition;
-            for (let i = 0; i < result.applied.length; i++) {
-                const move = result.applied[i];
-                const data = moveData?.[i];
-                const applyResult = applyMoveWithState(currentPos, move, {
-                    validateTurn: false,
-                });
-                if (applyResult.ok) {
-                    // 消費時間と評価値を渡す
-                    // KIFインポートの評価値は既に先手視点なので normalized: true
-                    navigation.addMove(move, applyResult.next, {
-                        elapsedMs: data?.elapsedMs,
-                        eval:
-                            data?.evalCp !== undefined || data?.evalMate !== undefined
-                                ? {
-                                      scoreCp: data.evalCp,
-                                      scoreMate: data.evalMate,
-                                      depth: data.depth,
-                                      normalized: true,
-                                  }
-                                : undefined,
-                    });
-                    currentPos = applyResult.next;
-                }
-            }
-
-            setLastMove(deriveLastMove(result.applied.at(-1)));
-            setSelection(null);
-            setMessage(null);
-            resetClocks(false);
-
-            clearLegalCache();
-            setPositionReady(true);
-
-            if (result.error) {
-                throw new Error(result.error);
-            }
-        },
-        [clearLegalCache, navigation, resetClocks, passRightsSettings],
-    );
-
-    // KIFコピー用コールバック
-    const handleCopyKif = useCallback((): string => {
-        return exportToKifString(kifMoves, boardHistory, {
-            startTime: new Date(),
-            senteName: sides.sente.role === "engine" ? "エンジン" : "人間",
-            goteName: sides.gote.role === "engine" ? "エンジン" : "人間",
-            includeEval: true, // 評価値もコメントとして出力
-            startSfen,
-        });
-    }, [kifMoves, boardHistory, sides.sente.role, sides.gote.role, startSfen]);
+    // 棋譜インポート・エクスポート管理フック
+    const { handleCopyKif, importSfen, importKif } = useKifuImportExport({
+        navigation,
+        passRightsSettings,
+        clearLegalCache,
+        resetClocks,
+        kifMoves,
+        boardHistory,
+        sides,
+        startSfen,
+        setLastMove,
+        setSelection,
+        setMessage,
+        setPositionReady,
+        setBasePosition,
+        setStartSfen,
+        setInitialBoard,
+        setLastAddedBranchInfo,
+        setIsEditMode,
+        setIsMatchRunning,
+    });
 
     // 棋譜の手数選択コールバック（巻き戻し・リプレイ用）
     const handlePlySelect = useCallback(
@@ -1876,96 +1820,6 @@ export function ShogiMatch({
             }
         },
         [],
-    );
-
-    // SFENインポート（局面 + 指し手）
-    // インポート後は自動的に検討モードに入る
-    const importSfen = useCallback(
-        async (sfen: string, movesToLoad: string[]) => {
-            const service = getPositionService();
-            try {
-                // 新しい開始局面を設定
-                const newPosition = await service.parseSfen(sfen);
-                setBasePosition(newPosition);
-                setStartSfen(sfen);
-                setInitialBoard(newPosition.board);
-
-                // 棋譜ナビゲーションをリセット
-                navigation.reset(newPosition, sfen);
-                setLastAddedBranchInfo(null); // 分岐状態をクリア
-
-                // 指し手がある場合は適用
-                if (movesToLoad.length > 0) {
-                    let currentPos = newPosition;
-                    const appliedMoves: string[] = [];
-                    for (const move of movesToLoad) {
-                        const applyResult = applyMoveWithState(currentPos, move, {
-                            validateTurn: false,
-                        });
-                        if (applyResult.ok) {
-                            navigation.addMove(move, applyResult.next);
-                            currentPos = applyResult.next;
-                            appliedMoves.push(move);
-                        } else {
-                            break;
-                        }
-                    }
-                    setLastMove(deriveLastMove(appliedMoves.at(-1)));
-                } else {
-                    setLastMove(undefined);
-                }
-
-                setSelection(null);
-                resetClocks(false);
-                clearLegalCache();
-                setPositionReady(true);
-
-                // インポート後は自動的に検討モードに入る
-                setIsEditMode(false);
-                setIsMatchRunning(false);
-            } catch (error) {
-                throw new Error(`SFENの適用に失敗しました: ${String(error)}`);
-            }
-        },
-        [clearLegalCache, navigation, resetClocks],
-    );
-
-    // KIFインポート（開始局面情報があれば使用）
-    // インポート後は自動的に検討モードに入る
-    const importKif = useCallback(
-        async (movesToLoad: string[], moveData: KifMoveData[], startSfenFromKif?: string) => {
-            const service = getPositionService();
-
-            let startPosition: PositionState;
-            let startSfenToLoad: string;
-
-            if (startSfenFromKif?.trim()) {
-                const parsed = parseSfen(startSfenFromKif);
-                if (!parsed.sfen) {
-                    throw new Error("開始局面のSFENが空です。");
-                }
-                startSfenToLoad = parsed.sfen;
-                try {
-                    startPosition = await service.parseSfen(startSfenToLoad);
-                } catch (error) {
-                    throw new Error(`開始局面の解析に失敗しました: ${String(error)}`);
-                }
-            } else {
-                startSfenToLoad = "startpos";
-                startPosition = await service.parseSfen(startSfenToLoad);
-            }
-
-            setBasePosition(startPosition);
-            setStartSfen(startSfenToLoad);
-            setInitialBoard(cloneBoard(startPosition.board));
-
-            await loadMoves(movesToLoad, moveData, startPosition, startSfenToLoad);
-
-            // KIFインポート後は自動的に検討モードに入る
-            setIsEditMode(false);
-            setIsMatchRunning(false);
-        },
-        [loadMoves],
     );
 
     const candidateNote = positionReady ? null : "局面を読み込み中です。";
